+++ src/main/java/bridge/BridgeGame.java
@@ -4,6 +4,7 @@ import bridge.domain.GameData;
 import bridge.dto.HistoryMapDto;
 import bridge.dto.PlayerLocationDto;
 import bridge.dto.PlayerTryCountDto;
+import bridge.exception.InputException;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
