+++ src/main/java/bridge/PlayerSetting.java
@@ -4,6 +4,7 @@ import bridge.domain.Player;
 import bridge.dto.BridgeCorrectLocationDto;
 import bridge.dto.BridgeDto;
 import bridge.dto.PlayerBridgeSizeDto;
+import bridge.exception.InputException;
 
 import java.util.List;
