+++ src/main/java/bridge/exception/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.exception;
 
 public enum ExceptionMessage {
     NOT_NUMBER("[ERROR] 숫자가 아님"),
