+++ src/main/java/bridge/exception/InputException.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.exception;
 
 import static bridge.PlayerCommend.*;
